pyproject.toml
[build-system]
requires = ["scikit-build-core>=0.5.0", "setuptools-scm"]
build-backend = "scikit_build_core.build"

[project]
name = "opengris-scaler"
description = "OpenGRIS Scaler Distribution Framework"
requires-python = ">=3.10"
readme = { file = "README.md", content-type = "text/markdown" }
license = { text = "Apache 2.0" }
authors = [{ name = "Citi", email = "opensource@citi.com" }]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: System :: Distributed Computing",
]
dynamic = ["version"]
dependencies = [
    "argcomplete",
    "bidict",
    "cloudpickle",
    "psutil==7.2.2",
    "pyzmq<27; python_version < '3.9'",
    "pyzmq; python_version >= '3.9'",
    "sortedcontainers==2.4.0",
    "tblib",
    "typing-extensions",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
uvloop = [
    "uvloop; platform_system != 'Windows'",
]
gui = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "websockets>=12.0",
]
graphblas = [
    "python-graphblas==2025.2.0",
    "numpy==2.2.6",
]
aws = [
    "boto3",
    "botocore[crt]",
]
orb = [
    "orb-py~=1.5.1; python_version >= '3.10'",
    "boto3",
    "packaging",
    "botocore[crt]",
]
all = [
    "opengris-scaler[aws]",
    # graphblas need compile on 3.14, and this python condition prevents up from building 3.14 wheel, comment it out for now
    # "opengris-scaler[graphblas]; python_version <= '3.13'",
    "opengris-scaler[gui]",
    "opengris-scaler[orb]",
    "opengris-scaler[uvloop]",
]

[dependency-groups]
dev = [
    "black>=24.8.0; python_version >= '3.10'",
    "isort>=7.0.0; python_version >= '3.10'",
    "flake8>=5.0.4",
    "flake8-pyproject>=1.2.3",
    "mypy>=1.14.1",
    "sphinx>=7.1.2",
    "websockets>=12.0",
]

[tool.scikit-build.metadata.version]
provider = "scikit_build_core.metadata.regex"
regex = '(?P<value>\d+\.\d+\.\d+(\.[1-9]\d*)?)'
input = "./src/scaler/version.txt"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.urls]
Home = "https://github.com/finos/opengris-scaler"

[project.scripts]
scaler_scheduler = "scaler.entry_points.scheduler:main"
scaler_top = "scaler.entry_points.top:main"
scaler_gui = "scaler.entry_points.webgui:main"
scaler_object_storage_server = "scaler.entry_points.object_storage_server:main"
scaler_worker_manager = "scaler.entry_points.worker_manager:main"
scaler = "scaler.entry_points.scaler:main"

[tool.scikit-build]
cmake.source-dir = "."
wheel.packages = ["src/scaler"]
build.targets = [
    "py_ymq",
    "py_object_storage_server",
    "py_one_to_many_dict",
    "py_stable_priority_queue",
    "py_many_to_many_dict",
    "py_indexed_queue",
    "py_capnp",
]

[tool.scikit-build.cmake.define]
CMAKE_C_COMPILER = { env = "CMAKE_C_COMPILER", default = "gcc" }
CMAKE_CXX_COMPILER = { env = "CMAKE_CXX_COMPILER", default = "g++" }

# WASM/Pyodide builds only ship the client-facing extensions.
# Skip the native compiler defaults so pyodide-build's emcc symlinks win.
[[tool.scikit-build.overrides]]
if.env.PYODIDE = true
build.targets = ["py_capnp"]
cmake.define.CMAKE_C_COMPILER = ""
cmake.define.CMAKE_CXX_COMPILER = ""

[tool.mypy]
no_strict_optional = true
check_untyped_defs = true
ignore_missing_imports = true
exclude = [
    "^docs.*$",
    "^benchmark.*$",
    "^venv.*$",
    "migrations",
    "__pycache__",
    "^build_*_*.*$",
    "^thirdparties.*$",
    "^capnproto.*$",
    "^boost.*$",
    "^scripts.*$",
    ".*object_storage.*$",
]

[tool.flake8]
count = true
statistics = true
max-line-length = 120
extend-ignore = ["E203"]
exclude = [
    "migrations",
    "__pycache__",
    "./build_*_*",
    "./thirdparties",
    "./capnproto*",
    "./boost*",
    "./scripts",
    "venv*"
]

[tool.black]
line-length = 120
skip-magic-trailing-comma = true
extend-exclude = '''
(
    ^/thirdparties
    | ^/build_.*
    | ^/capnproto.*
    | ^/boost.*
    | ^/scripts
    | ^/venv.*
)
'''

[tool.isort]
profile = "black"
line_length = 120
extend_skip_glob = [
    "thirdparties/*",
    "build_*",
    "capnproto*",
    "boost*",
    "scripts/*",
    "venv*",
]

[tool.cibuildwheel]
skip = "cp38-* cp39-*"

[tool.cibuildwheel.linux]
archs = ["x86_64", "aarch64"]
manylinux-x86_64-image = "manylinux_2_28"
manylinux-aarch64-image = "manylinux_2_28"
before-all = """
yum install -y cmake gcc-toolset-14-gcc gcc-toolset-14-gcc-c++; \
source /opt/rh/gcc-toolset-14/enable; \
./scripts/library_tool.sh capnp download; \
./scripts/library_tool.sh capnp compile; \
./scripts/library_tool.sh capnp install; \
./scripts/library_tool.sh libuv download; \
./scripts/library_tool.sh libuv compile; \
./scripts/library_tool.sh libuv install; \
echo "/usr/local/lib" > /etc/ld.so.conf.d/local.conf; \
ldconfig;
"""
environment = { PATH = "/opt/rh/gcc-toolset-14/root/usr/bin:$PATH", LD_LIBRARY_PATH = "/usr/local/lib:$LD_LIBRARY_PATH" }

[[tool.cibuildwheel.overrides]]
select = "*-musllinux*"
before-all = """
apk add cmake g++ make; \
./scripts/library_tool.sh capnp download; \
./scripts/library_tool.sh capnp compile; \
./scripts/library_tool.sh capnp install; \
./scripts/library_tool.sh libuv download; \
./scripts/library_tool.sh libuv compile; \
./scripts/library_tool.sh libuv install; \
ldconfig /usr/local/lib;
"""

[tool.cibuildwheel.macos]
archs = ["native"]
environment = { MACOSX_DEPLOYMENT_TARGET = "15.0" }
before-all = """
set -ex && \
./scripts/library_tool.sh capnp download; \
./scripts/library_tool.sh capnp compile --prefix="$CMAKE_PREFIX_PATH"; \
./scripts/library_tool.sh capnp install --prefix="$CMAKE_PREFIX_PATH"; \
./scripts/library_tool.sh libuv download; \
./scripts/library_tool.sh libuv compile --prefix="$CMAKE_PREFIX_PATH"; \
./scripts/library_tool.sh libuv install --prefix="$CMAKE_PREFIX_PATH"; \
"""

[tool.cibuildwheel.windows]
archs = ["AMD64"]
before-all = "pwsh -File ./scripts/cibuildwheel_before_all_windows.ps1"
environment = { CMAKE_C_COMPILER = "cl", CMAKE_CXX_COMPILER = "cl", CMAKE_GENERATOR = "Visual Studio 17 2022", CMAKE_GENERATOR_PLATFORM = "x64", CMAKE_CONFIGURATION_TYPES = "Release", CMAKE_PREFIX_PATH = "$TEMP/opengris-thirdparties" }
